formatter and revert in storage-postgresql tests
