+++ data-index/data-index-quarkus/data-index-storage-jpa-quarkus/src/test/java/org/kie/kogito/index/jpa/quarkus/query/H2ProcessInstanceEntityQueryIT.java
@@ -18,21 +18,21 @@
  */
 package org.kie.kogito.index.jpa.quarkus.query;
 
-import jakarta.inject.Inject;
 import org.kie.kogito.index.jpa.query.AbstractProcessInstanceEntityQueryIT;
+import org.kie.kogito.index.jpa.storage.ProcessInstanceEntityStorage;
 
 import io.quarkus.test.TestTransaction;
 import io.quarkus.test.junit.QuarkusTest;
-import org.kie.kogito.index.jpa.storage.ProcessDefinitionEntityStorage;
-import org.kie.kogito.index.jpa.storage.ProcessInstanceEntityStorage;
+
+import jakarta.inject.Inject;
 
 @QuarkusTest
 @TestTransaction
 class H2ProcessInstanceEntityQueryIT extends AbstractProcessInstanceEntityQueryIT {
 
     @Inject
-    public H2ProcessInstanceEntityQueryIT(ProcessInstanceEntityStorage storage, ProcessDefinitionEntityStorage definitionStorage) {
-        super(storage, definitionStorage);
+    public H2ProcessInstanceEntityQueryIT(ProcessInstanceEntityStorage storage) {
+        super(storage);
     }
 
     @Override
