+++ data-index/data-index-quarkus/data-index-storage-jpa-quarkus/src/test/java/org/kie/kogito/index/jpa/quarkus/query/H2UserTaskInstanceEntityQueryIT.java
@@ -18,12 +18,13 @@
  */
 package org.kie.kogito.index.jpa.quarkus.query;
 
-import jakarta.inject.Inject;
 import org.kie.kogito.index.jpa.query.AbstractUserTaskInstanceEntityQueryIT;
+import org.kie.kogito.index.jpa.storage.UserTaskInstanceEntityStorage;
 
 import io.quarkus.test.TestTransaction;
 import io.quarkus.test.junit.QuarkusTest;
-import org.kie.kogito.index.jpa.storage.UserTaskInstanceEntityStorage;
+
+import jakarta.inject.Inject;
 
 @QuarkusTest
 @TestTransaction
