+++ data-index/data-index-quarkus/data-index-storage-jpa-quarkus/src/test/java/org/kie/kogito/index/jpa/quarkus/storage/H2ProcessDefinitionStorageIT.java
@@ -18,12 +18,13 @@
  */
 package org.kie.kogito.index.jpa.quarkus.storage;
 
-import jakarta.inject.Inject;
 import org.kie.kogito.index.jpa.storage.AbstractProcessDefinitionStorageIT;
+import org.kie.kogito.index.jpa.storage.ProcessDefinitionEntityStorage;
 
 import io.quarkus.test.TestTransaction;
 import io.quarkus.test.junit.QuarkusTest;
-import org.kie.kogito.index.jpa.storage.ProcessDefinitionEntityStorage;
+
+import jakarta.inject.Inject;
 
 @QuarkusTest
 @TestTransaction
