+++ data-index/data-index-springboot/data-index-storage-jpa-springboot/src/test/java/org/kie/kogito/index/jpa/springboot/query/H2ProcessInstanceEntityQueryIT.java
@@ -20,7 +20,6 @@ package org.kie.kogito.index.jpa.springboot.query;
 
 import org.kie.kogito.index.jpa.query.AbstractProcessInstanceEntityQueryIT;
 import org.kie.kogito.index.jpa.springboot.KogitoSpringBootApplication;
-import org.kie.kogito.index.jpa.storage.ProcessDefinitionEntityStorage;
 import org.kie.kogito.index.jpa.storage.ProcessInstanceEntityStorage;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -33,8 +32,8 @@ import org.springframework.test.context.ActiveProfiles;
 class H2ProcessInstanceEntityQueryIT extends AbstractProcessInstanceEntityQueryIT {
 
     @Autowired
-    public H2ProcessInstanceEntityQueryIT(ProcessInstanceEntityStorage storage, ProcessDefinitionEntityStorage definitionStorage) {
-        super(storage, definitionStorage);
+    public H2ProcessInstanceEntityQueryIT(ProcessInstanceEntityStorage storage) {
+        super(storage);
     }
 
     @Override
