+++ data-index/data-index-storage/data-index-storage-jpa-common/src/test/java/org/kie/kogito/index/jpa/query/AbstractProcessDefinitionEntityQueryIT.java
@@ -19,7 +19,6 @@
 package org.kie.kogito.index.jpa.query;
 
 import java.util.List;
-import java.util.Map;
 import java.util.Set;
 
 import org.junit.jupiter.api.Test;
@@ -30,12 +29,7 @@ import org.kie.kogito.index.test.TestUtils;
 import org.kie.kogito.index.test.query.AbstractProcessDefinitionQueryIT;
 import org.kie.kogito.persistence.api.Storage;
 
-import static java.util.Collections.singletonList;
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.kie.kogito.index.json.JsonUtils.jsonFilter;
-import static org.kie.kogito.index.test.QueryTestUtils.assertNoKey;
-import static org.kie.kogito.index.test.QueryTestUtils.assertWithKey;
-import static org.kie.kogito.persistence.api.query.QueryFilterFactory.contains;
 import static org.kie.kogito.persistence.api.query.QueryFilterFactory.equalTo;
 
 public abstract class AbstractProcessDefinitionEntityQueryIT extends AbstractProcessDefinitionQueryIT {
@@ -58,20 +52,4 @@ public abstract class AbstractProcessDefinitionEntityQueryIT extends AbstractPro
         assertThat(storage.query().count()).isNotZero();
         assertThat(storage.query().filter(List.of(equalTo("version", "60.0"))).count()).isZero();
     }
-
-    @Test
-    void testMetadata() {
-        final String processId = "persons";
-        final String version = "1.0";
-        ProcessDefinitionKey key = new ProcessDefinitionKey(processId, version);
-        ProcessDefinition definitionEvent = TestUtils.createProcessDefinition(processId, version, Set.of());
-        definitionEvent.setMetadata(Map.of("name", "Javierito", "hobbies", List.of("community", "first")));
-        Storage<ProcessDefinitionKey, ProcessDefinition> storage = getStorage();
-        storage.put(key, definitionEvent);
-        queryAndAssert(assertWithKey(), storage, singletonList(jsonFilter(equalTo("metadata.name", "Javierito"))), null, null, null, key);
-        queryAndAssert(assertNoKey(), storage, singletonList(jsonFilter(equalTo("metadata.name", "Fulanito"))), null, null, null, key);
-        queryAndAssert(assertWithKey(), storage, singletonList(jsonFilter(contains("metadata.hobbies", "community"))), null, null, null, key);
-        queryAndAssert(assertNoKey(), storage, singletonList(jsonFilter(contains("metadata.hobbies", "commercial"))), null, null, null, key);
-    }
-
 }
