+++ data-index/data-index-storage/data-index-storage-jpa-common/src/test/java/org/kie/kogito/index/jpa/query/AbstractProcessInstanceEntityQueryIT.java
@@ -19,38 +19,24 @@
 package org.kie.kogito.index.jpa.query;
 
 import java.util.List;
-import java.util.Map;
-import java.util.Set;
 import java.util.UUID;
 
 import org.junit.jupiter.api.Test;
 import org.kie.kogito.event.process.ProcessInstanceStateDataEvent;
-import org.kie.kogito.event.process.ProcessInstanceVariableDataEvent;
-import org.kie.kogito.index.jpa.storage.ProcessDefinitionEntityStorage;
 import org.kie.kogito.index.jpa.storage.ProcessInstanceEntityStorage;
-import org.kie.kogito.index.model.ProcessDefinition;
-import org.kie.kogito.index.model.ProcessDefinitionKey;
-import org.kie.kogito.index.storage.ProcessInstanceStorage;
 import org.kie.kogito.index.test.TestUtils;
 import org.kie.kogito.index.test.query.AbstractProcessInstanceQueryIT;
 
-import static java.util.Collections.singletonList;
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.kie.kogito.index.json.JsonUtils.jsonFilter;
 import static org.kie.kogito.index.model.ProcessInstanceState.COMPLETED;
-import static org.kie.kogito.index.test.QueryTestUtils.assertNotId;
-import static org.kie.kogito.index.test.QueryTestUtils.assertWithId;
 import static org.kie.kogito.persistence.api.query.QueryFilterFactory.*;
-import static org.kie.kogito.persistence.api.query.QueryFilterFactory.contains;
 
 public abstract class AbstractProcessInstanceEntityQueryIT extends AbstractProcessInstanceQueryIT {
 
     ProcessInstanceEntityStorage storage;
-    ProcessDefinitionEntityStorage definitionStorage;
 
-    public AbstractProcessInstanceEntityQueryIT(ProcessInstanceEntityStorage storage, ProcessDefinitionEntityStorage definitionStorage) {
+    public AbstractProcessInstanceEntityQueryIT(ProcessInstanceEntityStorage storage) {
         this.storage = storage;
-        this.definitionStorage = definitionStorage;
     }
 
     @Override
@@ -64,102 +50,5 @@ public abstract class AbstractProcessInstanceEntityQueryIT extends AbstractProce
         storage.indexState(processInstanceEvent);
         assertThat(storage.query().count()).isNotZero();
         assertThat(storage.query().filter(List.of(in("state", List.of(34)))).count()).isZero();
-    }
-
-    @Test
-    void testProcessInstanceVariables() {
-        String processId = "travels";
-        String processInstanceId = UUID.randomUUID().toString();
-        ProcessInstanceStorage storage = getStorage();
-        ProcessInstanceVariableDataEvent variableEvent = TestUtils.createProcessInstanceVariableEvent(processInstanceId, processId, "John", 28, false,
-                List.of("Super", "Astonishing", "TheRealThing"));
-        final String version = "1.0";
-        ProcessDefinitionKey key = new ProcessDefinitionKey(processId, version);
-        ProcessDefinition definitionEvent = TestUtils.createProcessDefinition(processId, version, Set.of());
-        definitionEvent.setAnnotations(Set.of("Javierito", "Another"));
-        definitionEvent.setMetadata(Map.of("name", "Javierito", "hobbies", List.of("community", "first")));
-        variableEvent.setKogitoProcessInstanceVersion(version);
-        definitionStorage.put(key, definitionEvent);
-        storage.indexVariable(variableEvent);
-        queryAndAssert(assertWithId(), storage, singletonList(jsonFilter(equalTo("variables.traveller.name", "John"))), null, null, null,
-                processInstanceId);
-        queryAndAssert(assertNotId(), storage, singletonList(jsonFilter(equalTo("variables.traveller.name", "Smith"))), null, null, null,
-                processInstanceId);
-        queryAndAssert(assertWithId(), storage, singletonList(jsonFilter(equalTo("variables.traveller.isMartian", false))), null, null, null,
-                processInstanceId);
-        queryAndAssert(assertNotId(), storage, singletonList(jsonFilter(equalTo("variables.traveller.isMartian", true))), null, null, null,
-                processInstanceId);
-        queryAndAssert(assertWithId(), storage, singletonList(jsonFilter(equalTo("variables.traveller.age", 28))), null, null, null,
-                processInstanceId);
-        queryAndAssert(assertNotId(), storage, singletonList(jsonFilter(equalTo("variables.traveller.age", 29))), null, null, null,
-                processInstanceId);
-        queryAndAssert(assertWithId(), storage, singletonList(jsonFilter(between("variables.traveller.age", 26, 30))), null, null, null,
-                processInstanceId);
-        queryAndAssert(assertNotId(), storage, singletonList(jsonFilter(between("variables.traveller.age", 1, 3))), null, null, null,
-                processInstanceId);
-        queryAndAssert(assertWithId(), storage, singletonList(jsonFilter(between("variables.traveller.age", 26, 30))), null, null, null,
-                processInstanceId);
-        queryAndAssert(assertNotId(), storage, singletonList(jsonFilter(between("variables.traveller.age", 1, 3))), null, null, null,
-                processInstanceId);
-        queryAndAssert(assertWithId(), storage, singletonList(jsonFilter(greaterThan("variables.traveller.age", 26))), null, null, null,
-                processInstanceId);
-        queryAndAssert(assertNotId(), storage, singletonList(jsonFilter(greaterThan("variables.traveller.age", 28))), null, null, null,
-                processInstanceId);
-        queryAndAssert(assertWithId(), storage, singletonList(jsonFilter(greaterThanEqual("variables.traveller.age", 28))), null, null, null,
-                processInstanceId);
-        queryAndAssert(assertNotId(), storage, singletonList(jsonFilter(greaterThanEqual("variables.traveller.age", 29))), null, null, null,
-                processInstanceId);
-        queryAndAssert(assertWithId(), storage, singletonList(jsonFilter(lessThan("variables.traveller.age", 29))), null, null, null,
-                processInstanceId);
-        queryAndAssert(assertNotId(), storage, singletonList(jsonFilter(lessThan("variables.traveller.age", 28))), null, null, null,
-                processInstanceId);
-        queryAndAssert(assertWithId(), storage, singletonList(jsonFilter(lessThanEqual("variables.traveller.age", 28))), null, null, null,
-                processInstanceId);
-        queryAndAssert(assertNotId(), storage, singletonList(jsonFilter(lessThanEqual("variables.traveller.age", 27))), null, null, null,
-                processInstanceId);
-        queryAndAssert(assertWithId(), storage, singletonList(jsonFilter(lessThanEqual("variables.traveller.age", 28))), null, null, null,
-                processInstanceId);
-        queryAndAssert(assertNotId(), storage, singletonList(jsonFilter(lessThanEqual("variables.traveller.age", 27))), null, null, null,
-                processInstanceId);
-        queryAndAssert(assertWithId(), storage, singletonList(jsonFilter(in("variables.traveller.name", List.of("John", "Smith")))), null, null, null,
-                processInstanceId);
-        queryAndAssert(assertNotId(), storage, singletonList(jsonFilter(in("variables.traveller.age", List.of("Jack", "Smith")))), null, null, null,
-                processInstanceId);
-        queryAndAssert(assertWithId(), storage, singletonList(jsonFilter(in("variables.traveller.age", List.of(28, 29)))), null, null, null,
-                processInstanceId);
-        queryAndAssert(assertNotId(), storage, singletonList(jsonFilter(in("variables.traveller.age", List.of(27, 29)))), null, null, null,
-                processInstanceId);
-        queryAndAssert(assertWithId(), storage, singletonList(jsonFilter(like("variables.traveller.name", "Joh*"))), null, null, null,
-                processInstanceId);
-        queryAndAssert(assertNotId(), storage, singletonList(jsonFilter(like("variables.traveller.name", "Joha*"))), null, null, null,
-                processInstanceId);
-        queryAndAssert(assertWithId(), storage, singletonList(jsonFilter(notNull("variables.traveller.aliases"))), null, null, null,
-                processInstanceId);
-        queryAndAssert(assertNotId(), storage, singletonList(jsonFilter(isNull("variables.traveller.aliases"))), null, null, null,
-                processInstanceId);
-        queryAndAssert(assertWithId(), storage, singletonList(not(jsonFilter(isNull("variables.traveller.aliases")))), null, null, null,
-                processInstanceId);
-        queryAndAssert(assertWithId(), storage, singletonList(and(List.of(jsonFilter(notNull("variables.traveller.aliases")), jsonFilter(lessThan("variables.traveller.age", 45))))), null, null, null,
-                processInstanceId);
-        queryAndAssert(assertWithId(), storage, singletonList(or(List.of(jsonFilter(notNull("variables.traveller.aliases")), jsonFilter(lessThan("variables.traveller.age", 22))))), null, null, null,
-                processInstanceId);
-        queryAndAssert(assertWithId(), storage, singletonList(jsonFilter(contains("variables.traveller.aliases", "TheRealThing"))), null, null, null,
-                processInstanceId);
-        queryAndAssert(assertNotId(), storage, singletonList(jsonFilter(contains("variables.traveller.aliases", "TheDummyThing"))), null, null, null,
-                processInstanceId);
-        queryAndAssert(assertWithId(), storage, singletonList(jsonFilter(containsAny("variables.traveller.aliases", List.of("TheRealThing", "TheDummyThing")))), null, null, null,
-                processInstanceId);
-        queryAndAssert(assertNotId(), storage, singletonList(jsonFilter(containsAny("variables.traveller.aliases", List.of("TheRedPandaThing", "TheDummyThing")))), null, null, null,
-                processInstanceId);
-        queryAndAssert(assertWithId(), storage, singletonList(jsonFilter(containsAll("variables.traveller.aliases", List.of("Super", "Astonishing", "TheRealThing")))), null, null, null,
-                processInstanceId);
-        queryAndAssert(assertNotId(), storage, singletonList(jsonFilter(containsAll("variables.traveller.aliases", List.of("Super", "TheDummyThing")))), null, null, null,
-                processInstanceId);
-        queryAndAssert(assertWithId(), storage, singletonList(jsonFilter(equalTo("definition.metadata.name", "Javierito"))), null, null, null, processInstanceId);
-        queryAndAssert(assertNotId(), storage, singletonList(jsonFilter(equalTo("definition.metadata.name", "Fulanito"))), null, null, null, processInstanceId);
-        queryAndAssert(assertWithId(), storage, singletonList(jsonFilter(contains("definition.metadata.hobbies", "community"))), null, null, null, processInstanceId);
-        queryAndAssert(assertNotId(), storage, singletonList(jsonFilter(contains("definition.metadata.hobbies", "commercial"))), null, null, null, processInstanceId);
-        queryAndAssert(assertWithId(), storage, singletonList(contains("definition.annotations", "Javierito")), null, null, null, processInstanceId);
-        queryAndAssert(assertNotId(), storage, singletonList(contains("definition.annotations", "Fulanito")), null, null, null, processInstanceId);
     }
 }
