+++ data-index/data-index-storage/data-index-storage-postgresql/src/test/java/org/kie/kogito/index/postgresql/query/ProcessDefinitionEntityQueryIT.java
@@ -18,14 +18,30 @@
  */
 package org.kie.kogito.index.postgresql.query;
 
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
+
+import org.junit.jupiter.api.Test;
 import org.kie.kogito.index.jpa.query.AbstractProcessDefinitionEntityQueryIT;
 import org.kie.kogito.index.jpa.storage.ProcessDefinitionEntityStorage;
+import org.kie.kogito.index.model.ProcessDefinition;
+import org.kie.kogito.index.model.ProcessDefinitionKey;
+import org.kie.kogito.index.test.TestUtils;
+import org.kie.kogito.persistence.api.Storage;
 import org.kie.kogito.testcontainers.quarkus.PostgreSqlQuarkusTestResource;
 
 import io.quarkus.test.common.QuarkusTestResource;
 import io.quarkus.test.junit.QuarkusTest;
 
 import jakarta.inject.Inject;
+
+import static java.util.Collections.singletonList;
+import static org.kie.kogito.index.json.JsonUtils.jsonFilter;
+import static org.kie.kogito.index.test.QueryTestUtils.assertNoKey;
+import static org.kie.kogito.index.test.QueryTestUtils.assertWithKey;
+import static org.kie.kogito.persistence.api.query.QueryFilterFactory.contains;
+import static org.kie.kogito.persistence.api.query.QueryFilterFactory.equalTo;
 
 @QuarkusTest
 @QuarkusTestResource(PostgreSqlQuarkusTestResource.class)
@@ -34,5 +50,20 @@ class ProcessDefinitionEntityQueryIT extends AbstractProcessDefinitionEntityQuer
     @Inject
     public ProcessDefinitionEntityQueryIT(ProcessDefinitionEntityStorage storage) {
         super(storage);
+    }
+
+    @Test
+    void testMetadata() {
+        final String processId = "persons";
+        final String version = "1.0";
+        ProcessDefinitionKey key = new ProcessDefinitionKey(processId, version);
+        ProcessDefinition definitionEvent = TestUtils.createProcessDefinition(processId, version, Set.of());
+        definitionEvent.setMetadata(Map.of("name", "Javierito", "hobbies", List.of("community", "first")));
+        Storage<ProcessDefinitionKey, ProcessDefinition> storage = getStorage();
+        storage.put(key, definitionEvent);
+        queryAndAssert(assertWithKey(), storage, singletonList(jsonFilter(equalTo("metadata.name", "Javierito"))), null, null, null, key);
+        queryAndAssert(assertNoKey(), storage, singletonList(jsonFilter(equalTo("metadata.name", "Fulanito"))), null, null, null, key);
+        queryAndAssert(assertWithKey(), storage, singletonList(jsonFilter(contains("metadata.hobbies", "community"))), null, null, null, key);
+        queryAndAssert(assertNoKey(), storage, singletonList(jsonFilter(contains("metadata.hobbies", "commercial"))), null, null, null, key);
     }
 }
